Добавлена рефлексия и эндпоинты ответов

+++ flight-manager/src/main/java/y88/kirill/flight/manager/dto/request/roster/Assignment.java
@@ -12,7 +12,7 @@ public class Assignment{
     @JsonProperty("PairingIBMID") 
     public String pairingIBMID;
     @JsonProperty("PairingID") 
-    public int pairingID;
+    public long pairingID;
     @JsonProperty("Station") 
     public String station;
     @JsonProperty("DateTimeStart")
